Refactoring style and types
