+++ src/components/examples/column-pinning-table.tsx
