+++ src/components/examples/column-resizing-table.tsx
